updates experimental fragments
